requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==7.0
dnspython==2.3.0
Flask==1.1.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.12.0
python-dotenv==0.10.3
python-http-client==3.3.2
requests==2.25.1
soupsieve==2.3.1
urllib3==1.26.6
Werkzeug==1.0.1